+++ winwin-be-domain/src/main/java/com/dpm/winwin/domain/entity/member/MemberTalent.java
@@ -1,8 +1,7 @@
-package com.dpm.winwin.domain.entity.talent;
+package com.dpm.winwin.domain.entity.member;
 
 import com.dpm.winwin.domain.entity.BaseEntity;
 import com.dpm.winwin.domain.entity.category.SubCategory;
-import com.dpm.winwin.domain.entity.member.Member;
 import com.dpm.winwin.domain.entity.member.enums.TalentType;
 import lombok.AccessLevel;
 import lombok.Getter;
